xdr-agent/requirements.txt
psutil>=5.9.0
requests>=2.31.0
pyyaml>=6.0
apscheduler>=3.10.0
cryptography>=41.0.0
